libelf: Make all callers call elf_check_broken
This arranges that if the new pointer reference error checking
tripped, we actually get a message about it.  In this patch these
messages do not change the actual return values from the various
functions: so pointer reference errors do not prevent loading.  This
is for fear that some existing kernels might cause the code to make
these wild references, which would then break, which is not a good
thing in a security patch.

In xen/arch/x86/domain_build.c we have to introduce an "out" label and
change all of the "return rc" beyond the relevant point into "goto
out".

This is part of the fix to a security issue, XSA-55.

Differences in 4.1 backport:
 * No xen/arch/arm.
 * There was less error handling in xen/arch/x86/domain_build.c
   so less need to change it.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>