libxl: events: Makefile builds internal unit tests
We provide a new LIBXL_TESTS facility in the Makefile.

Each entry FOO in it has two main .c files,
  libxl_test_FOO.c  contains "inside libxl" code to support the test case
  test_FOO.c        contains "outside libxl" code to exercise the test case

We don't want to put the weird test case entrypoints and the weird
test case code in the main libxl.so library.  Symbol hiding prevents
us from simply directly linking the libxl_test_FOO.o in later.  So
instead we provide a special library libxenlight_test.so which is used
only locally.

So the upshot is that as far as the programmer is concerned the
"inside libxl" file is compiled exactly like a piece of libxl, and the
"outside libxl" file is compiled exactly like a piece of application
code.  They must share information via explicit libxl entrypoints.

We also provide some helpful common routines for unit tests to use.

There are not yet any test cases defined; that will come in the next
patch.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>