libxl: cancellation: Make spawns cancellable
The libxl__spawn_spawn internal API permits the caller to specify
.timeout_ms==-1, meaning to wait forever.  Provide an explicit
cancellable to allow spawns to be cancelled.

I think there are not currently any internal callers which do use
spawn with an infinite timeout, but this should not be left as a
lacuna for later.

Also, this change means that in practice anything which is cancelled
while spawning ought to be queued for two cancellation notifications:
one from its timeout, and one from the explicit cancellation
registration.  Hopefully this will mean that more exciting paths are
exercised during testing.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
---
v2: New in this version of the series.